move ProcessorTopologyTestDriver back to the test package
